Create several functions to add book to library
I created a template in the html for the book's card. Then, I created
several functions to contribute Title, Author, Pages, and Read values
to the DOM. Then, I used these functions in a larger
addBookToLibrary function. I used this function to add the template
pre-existing books for the library.