Add data-index-number attribute to book cards
I modified several parts of the code to allow data-indexing the created DOM elements.

I simplified the displayArray function to delete everything before displaying everything. This will allow the DOM to always be up to date with the current state of the Array.

I then added a setAttribute method to the displayCard() function that will check it's own index location in the array and set that number as it's data index number. This will be helpful when users delete books.

The updated displayArray function will allow for these deletions, as it will clear all children and re-display everything in the library area whenever it is called. This code will be called on the deletion function when it is created.

Lastly, I also organized the code and added/modified comments accordingly.